Refactor main slider and add debouncing.
This changes the way the slider works with doing a lot of
setActiveSlides and debouncing resize events. This also makes it so that
when resizing the side it should be completed after the slide resize is
complete.

Another change is that buttons are applied/deplied when the activeSlides
object changes. This is better than having to manually run it after each
resize or on initial load.