Try to get the last item without the width.
If we reach an item that goes over but its regular width still works we
continue. Once over it will break.

This should make it so that when a item is just over the viewport but
has some padding it will still be shown and the isLastSlide boolean is
set.