Bug fix gotoslide and undefined problem.
There was a issue where $viewport was not defined in time so we set
it up so coffee will evalulate each property as we need it.

Also fixed a issue where the goToSlide method on a single slide will
go past the last slide. So we made it so the index will remove one
slide from the index.